+++ smart_inventory_management/products/models.py
@@ -11,7 +11,7 @@ class Product(models.Model):
 
     cost_price = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
 
-    quantity = models.IntegerField()
+    quantity = models.DecimalField(max_digits=10, decimal_places=2)
 
     supplier = models.CharField(max_length=200)
